requirements.txt
asgiref==3.8.1
Django==5.1
djangorestframework==3.15.2
Faker==27.0.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
